# Minimize SentPacket size and allocations