`SentPacket` stores information about data transmitted in each packet.
This includes lists for all possible frame types.

However only a tiny subset of frame types are used in most packets:
Those will only contain ACKs and STREAM frames. The frame types
tracked in Retransmits make up less than 1/100 of all other types in
a common setup. E.g. the benchmark which transmits 1GB of data
shows:

```
frame_tx: FrameStats {
        ACK: 903061,
        CRYPTO: 2,
        NEW_CONNECTION_ID: 5,
        STREAM: 903057,
        // All other types are 0
}
```


We can exploit this property to only allocate tracking state for the
additional frame types if they are actually transmitted. This can be
done by boxing the `Retransmit` type in `SentPacket` and `SentFrames`.

Doing this reduces the sizes of the tracking data structures:

**Before**:

> Retransmits size: 184
> SentPacket size: 256

**After:**

> Retransmits size: 184
> SentPacket size: 80

This means memory consumption for packet tracking should be roughly
halved.

Besides this, the change avoids a mandatory allocation that currently
happens inside `Retransmits::default` for the `crypto: VecDeque`
list => In the best case we are avoiding one allocation per packet, in
the worst case the amount of allocations is the same.

The performance impact is very noticable - even without further
optimizations like GSO:

**Before:**

> Sent 1073741824 bytes on 1 streams in 5.11s (200.28 MiB/s)

**After:**

> Sent 1073741824 bytes on 1 streams in 4.80s (213.54 MiB/s)